Add project skeleton
About service
The purpose of the github-status service is to send messages about
incidents in Github components to VKTeams.
The service does not have a UI. It is assumed that the service
will be launched in the infrastructure and will run as a
background process inside the docker container.

The logic of the service
The service is based on an endless loop in which the following
actions are performed:
- With a specified interval (by default 30 seconds), an api
request is executed to check the status of Github components
- If an incident occurs in any component, information about it
is added to the internal incident registry
- If the incident is relevant for more than 2 minutes, a fail
message is sent to the VKTeams channel
- If all incidents are fixed, a message is sent that everything
is fixed

Part of #166